Back to using ports, with a little hacky reflection.
This attempt goes back to using ports, but without the need for the psasing
the port name as a string. This is pretty hacky, but instead we're parsing
the name of the incoming signal to figure out the port name.

I would love if Elm would provide the port and its conversion function as a first
class object rather only being on the runtime. In other words, I would love to
be able to pass a Port to the native JS.